Add Mask_Wire object
Add plotting functionality to all objects and importantly Layers.

Add requirements.txt to indicate the needed packages and versions to run the scripts.

Rename spuriously named "CleWin"-objects to be more generic, now "Mask".
More accurately indicate what is a cif-thing and what is a CleWin thing in code.

Document and comment on some code that needs to be better documented

Add wire example to example_with_alignment_marks

... and probably some more